Allow larger output sizes for Concat
UNet caused errors because the Concat operator could not handle
the large output sizes of some Concat operations.
The solution for this is to change the workgroup dimensions to
also use the y dimension.
Since there is a limit of threads per dimension, this allows
us quadratically increase the maximum possible output size.